Struct exif::experimental::Writer

pub struct Writer<'a> { /* fields omitted */ }

The Writer struct is used to encode and write Exif data.

Methods

impl<'a> Writer<'a>

pub fn new() -> Writer<'a>

Constructs an empty Writer.

pub fn push_field(&mut self, field: &'a Field)

Appends a field to be written.

The fields can be appended in any order. Duplicate fields must not be appended.

The following fields are ignored and synthesized when needed: ExifIFDPointer, GPSInfoIFDPointer, InteropIFDPointer, StripOffsets, StripByteCounts, TileOffsets, TileByteCounts, JPEGInterchangeFormat, and JPEGInterchangeFormatLength.

pub fn set_strips(&mut self, strips: &'a [&'a [u8]])

Sets TIFF strips for the primary image. If this method is called multiple times, the last one is used.

pub fn set_thumbnail_strips(&mut self, strips: &'a [&'a [u8]])

Sets TIFF strips for the thumbnail image. If this method is called multiple times, the last one is used.

pub fn set_tiles(&mut self, tiles: &'a [&'a [u8]])

Sets TIFF tiles for the primary image. If this method is called multiple times, the last one is used.

pub fn set_thumbnail_jpeg(&mut self, jpeg: &'a [u8])

Sets JPEG data for the thumbnail image. If this method is called multiple times, the last one is used.

pub fn write<W>(&mut self, w: &mut W, little_endian: bool) -> Result<(), Error> where
W: Write + Seek,

Encodes Exif data and writes it into w.

The write position of w must be set to zero before calling this method.