Struct exif::Reader

pub struct Reader { /* fields omitted */ }

The Reader struct reads a JPEG or TIFF image, parses the Exif attributes in it, and holds the results.

Methods

impl Reader

pub fn new<R>(reader: &mut R) -> Result<Reader, Error> where
R: BufRead,

Reads a JPEG or TIFF image and parses the Exif attributes in it. If an error occurred, exif::Error is returned.

pub fn buf(&self) -> &[u8]

Returns the slice that contains the TIFF data.

pub fn fields<'a>(&'a self) -> &[Field<'a>]

Returns a slice of Exif fields.

pub fn little_endian(&self) -> bool

Returns true if the TIFF data is in the little-endian byte order.

pub fn get_field(&self, tag: Tag, thumbnail: bool) -> Option<&Field>

Returns a reference to the Exif field specified by the tag and the thumbnail flag.