Struct exif::Tag

pub struct Tag(pub Context, pub u16);

A tag of a TIFF field.

Some well-known tags are provided as associated constants of this type. The constant names follow the Exif specification but not the Rust naming conventions.

Methods

impl Tag

pub fn context(self) -> Context

Returns the context of the tag.

pub fn number(self) -> u16

Returns the tag number.

pub fn description(&self) -> Option<&str>

Returns the description of the tag.

pub fn default_value(&self) -> Option<Value>

Returns the default value of the tag. None is returned if it is not defined in the standard or it depends on the context.

impl Tag

const ExifIFDPointer: Tag

ExifIFDPointer: Tag = Tag(Context::Tiff, 34665)

A pointer to the Exif IFD. This is used for the internal structure of Exif data and will not be returned to the user.

const GPSInfoIFDPointer: Tag

GPSInfoIFDPointer: Tag = Tag(Context::Tiff, 34853)

A pointer to the GPS IFD. This is used for the internal structure of Exif data and will not be returned to the user.

const InteropIFDPointer: Tag

InteropIFDPointer: Tag = Tag(Context::Exif, 40965)

A pointer to the interoperability IFD. This is used for the internal structure of Exif data and will not be returned to the user.

const ImageWidth: Tag

ImageWidth: Tag = Tag(Context::Tiff, 256)

const ImageLength: Tag

ImageLength: Tag = Tag(Context::Tiff, 257)

const BitsPerSample: Tag

BitsPerSample: Tag = Tag(Context::Tiff, 258)

const Compression: Tag

Compression: Tag = Tag(Context::Tiff, 259)

const PhotometricInterpretation: Tag

PhotometricInterpretation: Tag = Tag(Context::Tiff, 262)

const ImageDescription: Tag

ImageDescription: Tag = Tag(Context::Tiff, 270)

const Make: Tag

Make: Tag = Tag(Context::Tiff, 271)

const Model: Tag

Model: Tag = Tag(Context::Tiff, 272)

const StripOffsets: Tag

StripOffsets: Tag = Tag(Context::Tiff, 273)

const Orientation: Tag

Orientation: Tag = Tag(Context::Tiff, 274)

const SamplesPerPixel: Tag

SamplesPerPixel: Tag = Tag(Context::Tiff, 277)

const RowsPerStrip: Tag

RowsPerStrip: Tag = Tag(Context::Tiff, 278)

const StripByteCounts: Tag

StripByteCounts: Tag = Tag(Context::Tiff, 279)

const XResolution: Tag

XResolution: Tag = Tag(Context::Tiff, 282)

const YResolution: Tag

YResolution: Tag = Tag(Context::Tiff, 283)

const PlanarConfiguration: Tag

PlanarConfiguration: Tag = Tag(Context::Tiff, 284)

const ResolutionUnit: Tag

ResolutionUnit: Tag = Tag(Context::Tiff, 296)

const TransferFunction: Tag

TransferFunction: Tag = Tag(Context::Tiff, 301)

const Software: Tag

Software: Tag = Tag(Context::Tiff, 305)

const DateTime: Tag

DateTime: Tag = Tag(Context::Tiff, 306)

const Artist: Tag

Artist: Tag = Tag(Context::Tiff, 315)

const WhitePoint: Tag

WhitePoint: Tag = Tag(Context::Tiff, 318)

const PrimaryChromaticities: Tag

PrimaryChromaticities: Tag = Tag(Context::Tiff, 319)

const TileOffsets: Tag

TileOffsets: Tag = Tag(Context::Tiff, 324)

const TileByteCounts: Tag

TileByteCounts: Tag = Tag(Context::Tiff, 325)

const JPEGInterchangeFormat: Tag

JPEGInterchangeFormat: Tag = Tag(Context::Tiff, 513)

const JPEGInterchangeFormatLength: Tag

JPEGInterchangeFormatLength: Tag = Tag(Context::Tiff, 514)

const YCbCrCoefficients: Tag

YCbCrCoefficients: Tag = Tag(Context::Tiff, 529)

const YCbCrSubSampling: Tag

YCbCrSubSampling: Tag = Tag(Context::Tiff, 530)

const YCbCrPositioning: Tag

YCbCrPositioning: Tag = Tag(Context::Tiff, 531)

const ReferenceBlackWhite: Tag

ReferenceBlackWhite: Tag = Tag(Context::Tiff, 532)

const Copyright: Tag

Copyright: Tag = Tag(Context::Tiff, 33432)

const ExposureTime: Tag

ExposureTime: Tag = Tag(Context::Exif, 33434)

const FNumber: Tag

FNumber: Tag = Tag(Context::Exif, 33437)

const ExposureProgram: Tag

ExposureProgram: Tag = Tag(Context::Exif, 34850)

const SpectralSensitivity: Tag

SpectralSensitivity: Tag = Tag(Context::Exif, 34852)

const PhotographicSensitivity: Tag

PhotographicSensitivity: Tag = Tag(Context::Exif, 34855)

const OECF: Tag

OECF: Tag = Tag(Context::Exif, 34856)

const SensitivityType: Tag

SensitivityType: Tag = Tag(Context::Exif, 34864)

const StandardOutputSensitivity: Tag

StandardOutputSensitivity: Tag = Tag(Context::Exif, 34865)

const RecommendedExposureIndex: Tag

RecommendedExposureIndex: Tag = Tag(Context::Exif, 34866)

const ISOSpeed: Tag

ISOSpeed: Tag = Tag(Context::Exif, 34867)

const ISOSpeedLatitudeyyy: Tag

ISOSpeedLatitudeyyy: Tag = Tag(Context::Exif, 34868)

const ISOSpeedLatitudezzz: Tag

ISOSpeedLatitudezzz: Tag = Tag(Context::Exif, 34869)

const ExifVersion: Tag

ExifVersion: Tag = Tag(Context::Exif, 36864)

const DateTimeOriginal: Tag

DateTimeOriginal: Tag = Tag(Context::Exif, 36867)

const DateTimeDigitized: Tag

DateTimeDigitized: Tag = Tag(Context::Exif, 36868)

const OffsetTime: Tag

OffsetTime: Tag = Tag(Context::Exif, 36880)

const OffsetTimeOriginal: Tag

OffsetTimeOriginal: Tag = Tag(Context::Exif, 36881)

const OffsetTimeDigitized: Tag

OffsetTimeDigitized: Tag = Tag(Context::Exif, 36882)

const ComponentsConfiguration: Tag

ComponentsConfiguration: Tag = Tag(Context::Exif, 37121)

const CompressedBitsPerPixel: Tag

CompressedBitsPerPixel: Tag = Tag(Context::Exif, 37122)

const ShutterSpeedValue: Tag

ShutterSpeedValue: Tag = Tag(Context::Exif, 37377)

const ApertureValue: Tag

ApertureValue: Tag = Tag(Context::Exif, 37378)

const BrightnessValue: Tag

BrightnessValue: Tag = Tag(Context::Exif, 37379)

const ExposureBiasValue: Tag

ExposureBiasValue: Tag = Tag(Context::Exif, 37380)

const MaxApertureValue: Tag

MaxApertureValue: Tag = Tag(Context::Exif, 37381)

const SubjectDistance: Tag

SubjectDistance: Tag = Tag(Context::Exif, 37382)

const MeteringMode: Tag

MeteringMode: Tag = Tag(Context::Exif, 37383)

const LightSource: Tag

LightSource: Tag = Tag(Context::Exif, 37384)

const Flash: Tag

Flash: Tag = Tag(Context::Exif, 37385)

const FocalLength: Tag

FocalLength: Tag = Tag(Context::Exif, 37386)

const SubjectArea: Tag

SubjectArea: Tag = Tag(Context::Exif, 37396)

const MakerNote: Tag

MakerNote: Tag = Tag(Context::Exif, 37500)

const UserComment: Tag

UserComment: Tag = Tag(Context::Exif, 37510)

const SubSecTime: Tag

SubSecTime: Tag = Tag(Context::Exif, 37520)

const SubSecTimeOriginal: Tag

SubSecTimeOriginal: Tag = Tag(Context::Exif, 37521)

const SubSecTimeDigitized: Tag

SubSecTimeDigitized: Tag = Tag(Context::Exif, 37522)

const Temperature: Tag

Temperature: Tag = Tag(Context::Exif, 37888)

const Humidity: Tag

Humidity: Tag = Tag(Context::Exif, 37889)

const Pressure: Tag

Pressure: Tag = Tag(Context::Exif, 37890)

const WaterDepth: Tag

WaterDepth: Tag = Tag(Context::Exif, 37891)

const Acceleration: Tag

Acceleration: Tag = Tag(Context::Exif, 37892)

const CameraElevationAngle: Tag

CameraElevationAngle: Tag = Tag(Context::Exif, 37893)

const FlashpixVersion: Tag

FlashpixVersion: Tag = Tag(Context::Exif, 40960)

const ColorSpace: Tag

ColorSpace: Tag = Tag(Context::Exif, 40961)

const PixelXDimension: Tag

PixelXDimension: Tag = Tag(Context::Exif, 40962)

const PixelYDimension: Tag

PixelYDimension: Tag = Tag(Context::Exif, 40963)

const RelatedSoundFile: Tag

RelatedSoundFile: Tag = Tag(Context::Exif, 40964)

const FlashEnergy: Tag

FlashEnergy: Tag = Tag(Context::Exif, 41483)

const SpatialFrequencyResponse: Tag

SpatialFrequencyResponse: Tag = Tag(Context::Exif, 41484)

const FocalPlaneXResolution: Tag

FocalPlaneXResolution: Tag = Tag(Context::Exif, 41486)

const FocalPlaneYResolution: Tag

FocalPlaneYResolution: Tag = Tag(Context::Exif, 41487)

const FocalPlaneResolutionUnit: Tag

FocalPlaneResolutionUnit: Tag = Tag(Context::Exif, 41488)

const SubjectLocation: Tag

SubjectLocation: Tag = Tag(Context::Exif, 41492)

const ExposureIndex: Tag

ExposureIndex: Tag = Tag(Context::Exif, 41493)

const SensingMethod: Tag

SensingMethod: Tag = Tag(Context::Exif, 41495)

const FileSource: Tag

FileSource: Tag = Tag(Context::Exif, 41728)

const SceneType: Tag

SceneType: Tag = Tag(Context::Exif, 41729)

const CFAPattern: Tag

CFAPattern: Tag = Tag(Context::Exif, 41730)

const CustomRendered: Tag

CustomRendered: Tag = Tag(Context::Exif, 41985)

const ExposureMode: Tag

ExposureMode: Tag = Tag(Context::Exif, 41986)

const WhiteBalance: Tag

WhiteBalance: Tag = Tag(Context::Exif, 41987)

const DigitalZoomRatio: Tag

DigitalZoomRatio: Tag = Tag(Context::Exif, 41988)

const FocalLengthIn35mmFilm: Tag

FocalLengthIn35mmFilm: Tag = Tag(Context::Exif, 41989)

const SceneCaptureType: Tag

SceneCaptureType: Tag = Tag(Context::Exif, 41990)

const GainControl: Tag

GainControl: Tag = Tag(Context::Exif, 41991)

const Contrast: Tag

Contrast: Tag = Tag(Context::Exif, 41992)

const Saturation: Tag

Saturation: Tag = Tag(Context::Exif, 41993)

const Sharpness: Tag

Sharpness: Tag = Tag(Context::Exif, 41994)

const DeviceSettingDescription: Tag

DeviceSettingDescription: Tag = Tag(Context::Exif, 41995)

const SubjectDistanceRange: Tag

SubjectDistanceRange: Tag = Tag(Context::Exif, 41996)

const ImageUniqueID: Tag

ImageUniqueID: Tag = Tag(Context::Exif, 42016)

const CameraOwnerName: Tag

CameraOwnerName: Tag = Tag(Context::Exif, 42032)

const BodySerialNumber: Tag

BodySerialNumber: Tag = Tag(Context::Exif, 42033)

const LensSpecification: Tag

LensSpecification: Tag = Tag(Context::Exif, 42034)

const LensMake: Tag

LensMake: Tag = Tag(Context::Exif, 42035)

const LensModel: Tag

LensModel: Tag = Tag(Context::Exif, 42036)

const LensSerialNumber: Tag

LensSerialNumber: Tag = Tag(Context::Exif, 42037)

const Gamma: Tag

Gamma: Tag = Tag(Context::Exif, 42240)

const GPSVersionID: Tag

GPSVersionID: Tag = Tag(Context::Gps, 0)

const GPSLatitudeRef: Tag

GPSLatitudeRef: Tag = Tag(Context::Gps, 1)

const GPSLatitude: Tag

GPSLatitude: Tag = Tag(Context::Gps, 2)

const GPSLongitudeRef: Tag

GPSLongitudeRef: Tag = Tag(Context::Gps, 3)

const GPSLongitude: Tag

GPSLongitude: Tag = Tag(Context::Gps, 4)

const GPSAltitudeRef: Tag

GPSAltitudeRef: Tag = Tag(Context::Gps, 5)

const GPSAltitude: Tag

GPSAltitude: Tag = Tag(Context::Gps, 6)

const GPSTimeStamp: Tag

GPSTimeStamp: Tag = Tag(Context::Gps, 7)

const GPSSatellites: Tag

GPSSatellites: Tag = Tag(Context::Gps, 8)

const GPSStatus: Tag

GPSStatus: Tag = Tag(Context::Gps, 9)

const GPSMeasureMode: Tag

GPSMeasureMode: Tag = Tag(Context::Gps, 10)

const GPSDOP: Tag

GPSDOP: Tag = Tag(Context::Gps, 11)

const GPSSpeedRef: Tag

GPSSpeedRef: Tag = Tag(Context::Gps, 12)

const GPSSpeed: Tag

GPSSpeed: Tag = Tag(Context::Gps, 13)

const GPSTrackRef: Tag

GPSTrackRef: Tag = Tag(Context::Gps, 14)

const GPSTrack: Tag

GPSTrack: Tag = Tag(Context::Gps, 15)

const GPSImgDirectionRef: Tag

GPSImgDirectionRef: Tag = Tag(Context::Gps, 16)

const GPSImgDirection: Tag

GPSImgDirection: Tag = Tag(Context::Gps, 17)

const GPSMapDatum: Tag

GPSMapDatum: Tag = Tag(Context::Gps, 18)

const GPSDestLatitudeRef: Tag

GPSDestLatitudeRef: Tag = Tag(Context::Gps, 19)

const GPSDestLatitude: Tag

GPSDestLatitude: Tag = Tag(Context::Gps, 20)

const GPSDestLongitudeRef: Tag

GPSDestLongitudeRef: Tag = Tag(Context::Gps, 21)

const GPSDestLongitude: Tag

GPSDestLongitude: Tag = Tag(Context::Gps, 22)

const GPSDestBearingRef: Tag

GPSDestBearingRef: Tag = Tag(Context::Gps, 23)

const GPSDestBearing: Tag

GPSDestBearing: Tag = Tag(Context::Gps, 24)

const GPSDestDistanceRef: Tag

GPSDestDistanceRef: Tag = Tag(Context::Gps, 25)

const GPSDestDistance: Tag

GPSDestDistance: Tag = Tag(Context::Gps, 26)

const GPSProcessingMethod: Tag

GPSProcessingMethod: Tag = Tag(Context::Gps, 27)

const GPSAreaInformation: Tag

GPSAreaInformation: Tag = Tag(Context::Gps, 28)

const GPSDateStamp: Tag

GPSDateStamp: Tag = Tag(Context::Gps, 29)

const GPSDifferential: Tag

GPSDifferential: Tag = Tag(Context::Gps, 30)

const GPSHPositioningError: Tag

GPSHPositioningError: Tag = Tag(Context::Gps, 31)

const InteroperabilityIndex: Tag

InteroperabilityIndex: Tag = Tag(Context::Interop, 1)