Module exif::tag

A module that contains Exif tag constants.

Compatibility warning: Exif tag constants in this module will be converted to associated constants of Tag when the feature is stabilized.

It is not recommended to import the constants directly into your namespace; import the module and use with the module name like tag::DateTime. The constant names follow the Exif specification but not the Rust naming conventions, and a user of the constants will get the non_upper_case_globals warning if a bare constant is used in a match arm.

Constants

Acceleration 	 [
Deprecated
]

ApertureValue 	 [
Deprecated
]

Artist 	 [
Deprecated
]

BitsPerSample 	 [
Deprecated
]

BodySerialNumber 	 [
Deprecated
]

BrightnessValue 	 [
Deprecated
]

CFAPattern 	 [
Deprecated
]

CameraElevationAngle 	 [
Deprecated
]

CameraOwnerName 	 [
Deprecated
]

ColorSpace 	 [
Deprecated
]

ComponentsConfiguration 	 [
Deprecated
]

CompressedBitsPerPixel 	 [
Deprecated
]

Compression 	 [
Deprecated
]

Contrast 	 [
Deprecated
]

Copyright 	 [
Deprecated
]

CustomRendered 	 [
Deprecated
]

DateTime 	 [
Deprecated
]

DateTimeDigitized 	 [
Deprecated
]

DateTimeOriginal 	 [
Deprecated
]

DeviceSettingDescription 	 [
Deprecated
]

DigitalZoomRatio 	 [
Deprecated
]

ExifIFDPointer 	 [
Deprecated
]
A pointer to the Exif IFD. This is used for the internal structure of Exif data and will not be returned to the user.

ExifVersion 	 [
Deprecated
]

ExposureBiasValue 	 [
Deprecated
]

ExposureIndex 	 [
Deprecated
]

ExposureMode 	 [
Deprecated
]

ExposureProgram 	 [
Deprecated
]

ExposureTime 	 [
Deprecated
]

FNumber 	 [
Deprecated
]

FileSource 	 [
Deprecated
]

Flash 	 [
Deprecated
]

FlashEnergy 	 [
Deprecated
]

FlashpixVersion 	 [
Deprecated
]

FocalLength 	 [
Deprecated
]

FocalLengthIn35mmFilm 	 [
Deprecated
]

FocalPlaneResolutionUnit 	 [
Deprecated
]

FocalPlaneXResolution 	 [
Deprecated
]

FocalPlaneYResolution 	 [
Deprecated
]

GPSAltitude 	 [
Deprecated
]

GPSAltitudeRef 	 [
Deprecated
]

GPSAreaInformation 	 [
Deprecated
]

GPSDOP 	 [
Deprecated
]

GPSDateStamp 	 [
Deprecated
]

GPSDestBearing 	 [
Deprecated
]

GPSDestBearingRef 	 [
Deprecated
]

GPSDestDistance 	 [
Deprecated
]

GPSDestDistanceRef 	 [
Deprecated
]

GPSDestLatitude 	 [
Deprecated
]

GPSDestLatitudeRef 	 [
Deprecated
]

GPSDestLongitude 	 [
Deprecated
]

GPSDestLongitudeRef 	 [
Deprecated
]

GPSDifferential 	 [
Deprecated
]

GPSHPositioningError 	 [
Deprecated
]

GPSImgDirection 	 [
Deprecated
]

GPSImgDirectionRef 	 [
Deprecated
]

GPSInfoIFDPointer 	 [
Deprecated
]
A pointer to the GPS IFD. This is used for the internal structure of Exif data and will not be returned to the user.

GPSLatitude 	 [
Deprecated
]

GPSLatitudeRef 	 [
Deprecated
]

GPSLongitude 	 [
Deprecated
]

GPSLongitudeRef 	 [
Deprecated
]

GPSMapDatum 	 [
Deprecated
]

GPSMeasureMode 	 [
Deprecated
]

GPSProcessingMethod 	 [
Deprecated
]

GPSSatellites 	 [
Deprecated
]

GPSSpeed 	 [
Deprecated
]

GPSSpeedRef 	 [
Deprecated
]

GPSStatus 	 [
Deprecated
]

GPSTimeStamp 	 [
Deprecated
]

GPSTrack 	 [
Deprecated
]

GPSTrackRef 	 [
Deprecated
]

GPSVersionID 	 [
Deprecated
]

GainControl 	 [
Deprecated
]

Gamma 	 [
Deprecated
]

Humidity 	 [
Deprecated
]

ISOSpeed 	 [
Deprecated
]

ISOSpeedLatitudeyyy 	 [
Deprecated
]

ISOSpeedLatitudezzz 	 [
Deprecated
]

ImageDescription 	 [
Deprecated
]

ImageLength 	 [
Deprecated
]

ImageUniqueID 	 [
Deprecated
]

ImageWidth 	 [
Deprecated
]

InteropIFDPointer 	 [
Deprecated
]
A pointer to the interoperability IFD. This is used for the internal structure of Exif data and will not be returned to the user.

InteroperabilityIndex 	 [
Deprecated
]

JPEGInterchangeFormat 	 [
Deprecated
]

JPEGInterchangeFormatLength 	 [
Deprecated
]

LensMake 	 [
Deprecated
]

LensModel 	 [
Deprecated
]

LensSerialNumber 	 [
Deprecated
]

LensSpecification 	 [
Deprecated
]

LightSource 	 [
Deprecated
]

Make 	 [
Deprecated
]

MakerNote 	 [
Deprecated
]

MaxApertureValue 	 [
Deprecated
]

MeteringMode 	 [
Deprecated
]

Model 	 [
Deprecated
]

OECF 	 [
Deprecated
]

OffsetTime 	 [
Deprecated
]

OffsetTimeDigitized 	 [
Deprecated
]

OffsetTimeOriginal 	 [
Deprecated
]

Orientation 	 [
Deprecated
]

PhotographicSensitivity 	 [
Deprecated
]

PhotometricInterpretation 	 [
Deprecated
]

PixelXDimension 	 [
Deprecated
]

PixelYDimension 	 [
Deprecated
]

PlanarConfiguration 	 [
Deprecated
]

Pressure 	 [
Deprecated
]

PrimaryChromaticities 	 [
Deprecated
]

RecommendedExposureIndex 	 [
Deprecated
]

ReferenceBlackWhite 	 [
Deprecated
]

RelatedSoundFile 	 [
Deprecated
]

ResolutionUnit 	 [
Deprecated
]

RowsPerStrip 	 [
Deprecated
]

SamplesPerPixel 	 [
Deprecated
]

Saturation 	 [
Deprecated
]

SceneCaptureType 	 [
Deprecated
]

SceneType 	 [
Deprecated
]

SensingMethod 	 [
Deprecated
]

SensitivityType 	 [
Deprecated
]

Sharpness 	 [
Deprecated
]

ShutterSpeedValue 	 [
Deprecated
]

Software 	 [
Deprecated
]

SpatialFrequencyResponse 	 [
Deprecated
]

SpectralSensitivity 	 [
Deprecated
]

StandardOutputSensitivity 	 [
Deprecated
]

StripByteCounts 	 [
Deprecated
]

StripOffsets 	 [
Deprecated
]

SubSecTime 	 [
Deprecated
]

SubSecTimeDigitized 	 [
Deprecated
]

SubSecTimeOriginal 	 [
Deprecated
]

SubjectArea 	 [
Deprecated
]

SubjectDistance 	 [
Deprecated
]

SubjectDistanceRange 	 [
Deprecated
]

SubjectLocation 	 [
Deprecated
]

Temperature 	 [
Deprecated
]

TileByteCounts 	 [
Deprecated
]

TileOffsets 	 [
Deprecated
]

TransferFunction 	 [
Deprecated
]

UserComment 	 [
Deprecated
]

WaterDepth 	 [
Deprecated
]

WhiteBalance 	 [
Deprecated
]

WhitePoint 	 [
Deprecated
]

XResolution 	 [
Deprecated
]

YCbCrCoefficients 	 [
Deprecated
]

YCbCrPositioning 	 [
Deprecated
]

YCbCrSubSampling 	 [
Deprecated
]

YResolution 	 [
Deprecated
]