exif

Struct Tag

pub struct Tag(pub Context, pub u16);

A tag of a TIFF/Exif field.

Some well-known tags are provided as associated constants of this type. The constant names follow the Exif specification but not the Rust naming conventions.

A non-predefined tag can also be specified by the context and the number as in Tag(Context::Tiff, 0x100).

Tuple Fields
0: Context1: u16
Implementations

impl Tag

pub fn context(self) -> Context

Returns the context of the tag.

Examples

use exif::{Context, Tag};
assert_eq!(Tag::XResolution.context(), Context::Tiff);
assert_eq!(Tag::DateTimeOriginal.context(), Context::Exif);

pub fn number(self) -> u16

Returns the tag number.

Examples

use exif::Tag;
assert_eq!(Tag::XResolution.number(), 0x11a);
assert_eq!(Tag::DateTimeOriginal.number(), 0x9003);

pub fn description(&self) -> Option<&str>

Returns the description of the tag.

pub fn default_value(&self) -> Option<Value>

Returns the default value of the tag. None is returned if it is not defined in the standard or it depends on another tag.

impl Tag

pub const ImageWidth: Tag

pub const ImageLength: Tag

pub const BitsPerSample: Tag

pub const Compression: Tag

pub const PhotometricInterpretation: Tag

pub const ImageDescription: Tag

pub const Make: Tag

Manufacturer of the image input equipment.

pub const Model: Tag

Model name or number of the image input equipment.

pub const StripOffsets: Tag

pub const Orientation: Tag

pub const SamplesPerPixel: Tag

pub const RowsPerStrip: Tag

pub const StripByteCounts: Tag

pub const XResolution: Tag

pub const YResolution: Tag

pub const PlanarConfiguration: Tag

pub const ResolutionUnit: Tag

Unit of XResolution and YResolution fields.

pub const TransferFunction: Tag

pub const Software: Tag

Name and version of the software used to create the image.

pub const DateTime: Tag

Date and time when the image file was created or last edited. For the time when the picture was taken, see DateTimeOriginal field.

pub const Artist: Tag

pub const WhitePoint: Tag

pub const PrimaryChromaticities: Tag

pub const TileOffsets: Tag

pub const TileByteCounts: Tag

pub const JPEGInterchangeFormat: Tag

pub const JPEGInterchangeFormatLength: Tag

pub const YCbCrCoefficients: Tag

pub const YCbCrSubSampling: Tag

pub const YCbCrPositioning: Tag

pub const ReferenceBlackWhite: Tag

pub const Copyright: Tag

pub const ExposureTime: Tag

pub const FNumber: Tag

pub const ExposureProgram: Tag

pub const SpectralSensitivity: Tag

pub const PhotographicSensitivity: Tag

Sensitivity of the device.

The value may be represented by standard output sensitivity (SOS), recommended exposure index (REI), or ISO speed. What is stored is designated by SensitivityType field.

This field is 16-bit and may be saturated. For 32-bit values, see StandardOutputSensitivity, RecommendedExposureIndex, ISOSpeed, ISOSpeedLatitudeyyy, and ISOSpeedLatitudezzz fields.

pub const OECF: Tag

pub const SensitivityType: Tag

pub const StandardOutputSensitivity: Tag

pub const RecommendedExposureIndex: Tag

pub const ISOSpeed: Tag

pub const ISOSpeedLatitudeyyy: Tag

pub const ISOSpeedLatitudezzz: Tag

pub const ExifVersion: Tag

pub const DateTimeOriginal: Tag

Date and time when the original image was generated (e.g., the picture was taken by a camera).

pub const DateTimeDigitized: Tag

Date and time when the image was stored as digital data. If a picture is taken by a film camera and then digitized later, this value will be different from DateTimeOriginal field.

pub const OffsetTime: Tag

Timezone offset for DateTime field.

pub const OffsetTimeOriginal: Tag

Timezone offset for DateTimeOriginal field.

pub const OffsetTimeDigitized: Tag

Timezone offset for DateTimeDigitized field.

pub const ComponentsConfiguration: Tag

pub const CompressedBitsPerPixel: Tag

pub const ShutterSpeedValue: Tag

pub const ApertureValue: Tag

pub const BrightnessValue: Tag

pub const ExposureBiasValue: Tag

pub const MaxApertureValue: Tag

pub const SubjectDistance: Tag

pub const MeteringMode: Tag

pub const LightSource: Tag

pub const Flash: Tag

pub const FocalLength: Tag

pub const SubjectArea: Tag

pub const MakerNote: Tag

pub const UserComment: Tag

pub const SubSecTime: Tag

Subseconds for DateTime field.

pub const SubSecTimeOriginal: Tag

Subseconds for DateTimeOriginal field.

pub const SubSecTimeDigitized: Tag

Subseconds for DateTimeDigitized field.

pub const Temperature: Tag

pub const Humidity: Tag

pub const Pressure: Tag

pub const WaterDepth: Tag

pub const Acceleration: Tag

pub const CameraElevationAngle: Tag

pub const FlashpixVersion: Tag

pub const ColorSpace: Tag

pub const PixelXDimension: Tag

pub const PixelYDimension: Tag

pub const RelatedSoundFile: Tag

pub const FlashEnergy: Tag

pub const SpatialFrequencyResponse: Tag

pub const FocalPlaneXResolution: Tag

pub const FocalPlaneYResolution: Tag

pub const FocalPlaneResolutionUnit: Tag

Unit of FocalPlaneXResolution and FocalPlaneYResolution fields.

pub const SubjectLocation: Tag

pub const ExposureIndex: Tag

pub const SensingMethod: Tag

pub const FileSource: Tag

pub const SceneType: Tag

pub const CFAPattern: Tag

pub const CustomRendered: Tag

pub const ExposureMode: Tag

pub const WhiteBalance: Tag

pub const DigitalZoomRatio: Tag

pub const FocalLengthIn35mmFilm: Tag

pub const SceneCaptureType: Tag

pub const GainControl: Tag

pub const Contrast: Tag

pub const Saturation: Tag

pub const Sharpness: Tag

pub const DeviceSettingDescription: Tag

pub const SubjectDistanceRange: Tag

pub const ImageUniqueID: Tag

pub const CameraOwnerName: Tag

pub const BodySerialNumber: Tag

pub const LensSpecification: Tag

pub const LensMake: Tag

pub const LensModel: Tag

pub const LensSerialNumber: Tag

pub const CompositeImage: Tag

pub const SourceImageNumberOfCompositeImage: Tag

pub const SourceExposureTimesOfCompositeImage: Tag

pub const Gamma: Tag

pub const GPSVersionID: Tag

pub const GPSLatitudeRef: Tag

pub const GPSLatitude: Tag

pub const GPSLongitudeRef: Tag

pub const GPSLongitude: Tag

pub const GPSAltitudeRef: Tag

pub const GPSAltitude: Tag

pub const GPSTimeStamp: Tag

pub const GPSSatellites: Tag

pub const GPSStatus: Tag

pub const GPSMeasureMode: Tag

pub const GPSDOP: Tag

pub const GPSSpeedRef: Tag

pub const GPSSpeed: Tag

pub const GPSTrackRef: Tag

pub const GPSTrack: Tag

pub const GPSImgDirectionRef: Tag

pub const GPSImgDirection: Tag

pub const GPSMapDatum: Tag

pub const GPSDestLatitudeRef: Tag

pub const GPSDestLatitude: Tag

pub const GPSDestLongitudeRef: Tag

pub const GPSDestLongitude: Tag

pub const GPSDestBearingRef: Tag

pub const GPSDestBearing: Tag

pub const GPSDestDistanceRef: Tag

pub const GPSDestDistance: Tag

pub const GPSProcessingMethod: Tag

pub const GPSAreaInformation: Tag

pub const GPSDateStamp: Tag

pub const GPSDifferential: Tag

pub const GPSHPositioningError: Tag

pub const InteroperabilityIndex: Tag

pub const InteroperabilityVersion: Tag

pub const RelatedImageFileFormat: Tag

pub const RelatedImageWidth: Tag

pub const RelatedImageLength: Tag