Struct exif::Tag

pub struct Tag(pub Context, pub u16);

A tag of a TIFF/Exif field.

Some well-known tags are provided as associated constants of this type. The constant names follow the Exif specification but not the Rust naming conventions.

A non-predefined tag can also be specified by the context and the number as in Tag(Context::Tiff, 0x100).

Tuple Fields
0: Context1: u16
Implementations

impl Tag

pub fn context(self) -> Context

Returns the context of the tag.

Examples

use exif::{Context, Tag};
assert_eq!(Tag::XResolution.context(), Context::Tiff);
assert_eq!(Tag::DateTimeOriginal.context(), Context::Exif);

pub fn number(self) -> u16

Returns the tag number.

Examples

use exif::Tag;
assert_eq!(Tag::XResolution.number(), 0x11a);
assert_eq!(Tag::DateTimeOriginal.number(), 0x9003);

pub fn description(&self) -> Option<&str>

Returns the description of the tag.

pub fn default_value(&self) -> Option<Value>

Returns the default value of the tag. None is returned if it is not defined in the standard or it depends on another tag.

impl Tag

pub const ImageWidth: Tag = _

pub const ImageLength: Tag = _

pub const BitsPerSample: Tag = _

pub const Compression: Tag = _

pub const PhotometricInterpretation: Tag = _

pub const ImageDescription: Tag = _

pub const Make: Tag = _

Manufacturer of the image input equipment.

pub const Model: Tag = _

Model name or number of the image input equipment.

pub const StripOffsets: Tag = _

pub const Orientation: Tag = _

pub const SamplesPerPixel: Tag = _

pub const RowsPerStrip: Tag = _

pub const StripByteCounts: Tag = _

pub const XResolution: Tag = _

pub const YResolution: Tag = _

pub const PlanarConfiguration: Tag = _

pub const ResolutionUnit: Tag = _

Unit of XResolution and YResolution fields.

pub const TransferFunction: Tag = _

pub const Software: Tag = _

Name and version of the software used to create the image.

pub const DateTime: Tag = _

Date and time when the image file was created or last edited. For the time when the picture was taken, see DateTimeOriginal field.

pub const Artist: Tag = _

pub const WhitePoint: Tag = _

pub const PrimaryChromaticities: Tag = _

pub const TileOffsets: Tag = _

pub const TileByteCounts: Tag = _

pub const JPEGInterchangeFormat: Tag = _

pub const JPEGInterchangeFormatLength: Tag = _

pub const YCbCrCoefficients: Tag = _

pub const YCbCrSubSampling: Tag = _

pub const YCbCrPositioning: Tag = _

pub const ReferenceBlackWhite: Tag = _

pub const Copyright: Tag = _

pub const ExposureTime: Tag = _

pub const FNumber: Tag = _

pub const ExposureProgram: Tag = _

pub const SpectralSensitivity: Tag = _

pub const PhotographicSensitivity: Tag = _

Sensitivity of the device.

The value may be represented by standard output sensitivity (SOS), recommended exposure index (REI), or ISO speed. What is stored is designated by SensitivityType field.

This field is 16-bit and may be saturated. For 32-bit values, see StandardOutputSensitivity, RecommendedExposureIndex, ISOSpeed, ISOSpeedLatitudeyyy, and ISOSpeedLatitudezzz fields.

pub const OECF: Tag = _

pub const SensitivityType: Tag = _

pub const StandardOutputSensitivity: Tag = _

pub const RecommendedExposureIndex: Tag = _

pub const ISOSpeed: Tag = _

pub const ISOSpeedLatitudeyyy: Tag = _

pub const ISOSpeedLatitudezzz: Tag = _

pub const ExifVersion: Tag = _

pub const DateTimeOriginal: Tag = _

Date and time when the original image was generated (e.g., the picture was taken by a camera).

pub const DateTimeDigitized: Tag = _

Date and time when the image was stored as digital data. If a picture is taken by a film camera and then digitized later, this value will be different from DateTimeOriginal field.

pub const OffsetTime: Tag = _

Timezone offset for DateTime field.

pub const OffsetTimeOriginal: Tag = _

Timezone offset for DateTimeOriginal field.

pub const OffsetTimeDigitized: Tag = _

Timezone offset for DateTimeDigitized field.

pub const ComponentsConfiguration: Tag = _

pub const CompressedBitsPerPixel: Tag = _

pub const ShutterSpeedValue: Tag = _

pub const ApertureValue: Tag = _

pub const BrightnessValue: Tag = _

pub const ExposureBiasValue: Tag = _

pub const MaxApertureValue: Tag = _

pub const SubjectDistance: Tag = _

pub const MeteringMode: Tag = _

pub const LightSource: Tag = _

pub const Flash: Tag = _

pub const FocalLength: Tag = _

pub const SubjectArea: Tag = _

pub const MakerNote: Tag = _

pub const UserComment: Tag = _

pub const SubSecTime: Tag = _

Subseconds for DateTime field.

pub const SubSecTimeOriginal: Tag = _

Subseconds for DateTimeOriginal field.

pub const SubSecTimeDigitized: Tag = _

Subseconds for DateTimeDigitized field.

pub const Temperature: Tag = _

pub const Humidity: Tag = _

pub const Pressure: Tag = _

pub const WaterDepth: Tag = _

pub const Acceleration: Tag = _

pub const CameraElevationAngle: Tag = _

pub const FlashpixVersion: Tag = _

pub const ColorSpace: Tag = _

pub const PixelXDimension: Tag = _

pub const PixelYDimension: Tag = _

pub const RelatedSoundFile: Tag = _

pub const FlashEnergy: Tag = _

pub const SpatialFrequencyResponse: Tag = _

pub const FocalPlaneXResolution: Tag = _

pub const FocalPlaneYResolution: Tag = _

pub const FocalPlaneResolutionUnit: Tag = _

Unit of FocalPlaneXResolution and FocalPlaneYResolution fields.

pub const SubjectLocation: Tag = _

pub const ExposureIndex: Tag = _

pub const SensingMethod: Tag = _

pub const FileSource: Tag = _

pub const SceneType: Tag = _

pub const CFAPattern: Tag = _

pub const CustomRendered: Tag = _

pub const ExposureMode: Tag = _

pub const WhiteBalance: Tag = _

pub const DigitalZoomRatio: Tag = _

pub const FocalLengthIn35mmFilm: Tag = _

pub const SceneCaptureType: Tag = _

pub const GainControl: Tag = _

pub const Contrast: Tag = _

pub const Saturation: Tag = _

pub const Sharpness: Tag = _

pub const DeviceSettingDescription: Tag = _

pub const SubjectDistanceRange: Tag = _

pub const ImageUniqueID: Tag = _

pub const CameraOwnerName: Tag = _

pub const BodySerialNumber: Tag = _

pub const LensSpecification: Tag = _

pub const LensMake: Tag = _

pub const LensModel: Tag = _

pub const LensSerialNumber: Tag = _

pub const CompositeImage: Tag = _

pub const SourceImageNumberOfCompositeImage: Tag = _

pub const SourceExposureTimesOfCompositeImage: Tag = _

pub const Gamma: Tag = _

pub const GPSVersionID: Tag = _

pub const GPSLatitudeRef: Tag = _

pub const GPSLatitude: Tag = _

pub const GPSLongitudeRef: Tag = _

pub const GPSLongitude: Tag = _

pub const GPSAltitudeRef: Tag = _

pub const GPSAltitude: Tag = _

pub const GPSTimeStamp: Tag = _

pub const GPSSatellites: Tag = _

pub const GPSStatus: Tag = _

pub const GPSMeasureMode: Tag = _

pub const GPSDOP: Tag = _

pub const GPSSpeedRef: Tag = _

pub const GPSSpeed: Tag = _

pub const GPSTrackRef: Tag = _

pub const GPSTrack: Tag = _

pub const GPSImgDirectionRef: Tag = _

pub const GPSImgDirection: Tag = _

pub const GPSMapDatum: Tag = _

pub const GPSDestLatitudeRef: Tag = _

pub const GPSDestLatitude: Tag = _

pub const GPSDestLongitudeRef: Tag = _

pub const GPSDestLongitude: Tag = _

pub const GPSDestBearingRef: Tag = _

pub const GPSDestBearing: Tag = _

pub const GPSDestDistanceRef: Tag = _

pub const GPSDestDistance: Tag = _

pub const GPSProcessingMethod: Tag = _

pub const GPSAreaInformation: Tag = _

pub const GPSDateStamp: Tag = _

pub const GPSDifferential: Tag = _

pub const GPSHPositioningError: Tag = _

pub const InteroperabilityIndex: Tag = _

pub const InteroperabilityVersion: Tag = _

pub const RelatedImageFileFormat: Tag = _

pub const RelatedImageWidth: Tag = _

pub const RelatedImageLength: Tag = _